+++ src/Profile/Magento/DataSelection/ProductDataSelection.php
@@ -7,6 +7,7 @@
 
 namespace Swag\MigrationMagento\Profile\Magento\DataSelection;
 
+use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\CrossSellingDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ManufacturerDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductCustomFieldDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductDataSet;
@@ -44,6 +45,7 @@ class ProductDataSelection implements DataSelectionInterface
             PropertyGroupDataSet::getEntity(),
             ProductCustomFieldDataSet::getEntity(),
             ProductDataSet::getEntity(),
+            CrossSellingDataSet::getEntity(),
         ];
     }