+++ tests/_fixtures/cross_selling_data.php
@@ -1,0 +1,58 @@
+<?php declare(strict_types=1);
+/*
+ * (c) shopware AG <info@shopware.com>
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+return [
+    [
+        'link_id' => '632',
+        'product_id' => '878',
+        'linked_product_id' => '892',
+        'link_type_id' => '4',
+        'type' => 'up_sell',
+        'sourceProductId' => '877',
+        'position' => 16,
+    ],
+
+    [
+        'link_id' => '633',
+        'product_id' => '878',
+        'linked_product_id' => '807',
+        'link_type_id' => '5',
+        'type' => 'cross_sell',
+        'sourceProductId' => '882',
+        'position' => 17,
+    ],
+
+    [
+        'link_id' => '629',
+        'product_id' => '881',
+        'linked_product_id' => '879',
+        'link_type_id' => '5',
+        'type' => 'cross_sell',
+        'sourceProductId' => '877',
+        'position' => 11,
+    ],
+
+    [
+        'link_id' => '632',
+        'product_id' => '878',
+        'linked_product_id' => '892',
+        'link_type_id' => '4',
+        'type' => 'up_sell',
+        'sourceProductId' => '417',
+        'position' => 6,
+    ],
+
+    [
+        'link_id' => '636',
+        'product_id' => '878',
+        'linked_product_id' => '895',
+        'link_type_id' => '1',
+        'type' => 'relation',
+        'sourceProductId' => '877',
+        'position' => 18,
+    ],
+];